BspUart重构，统一其他调用的地方
