+++ ReactorLibs/Bsps/bsp_system/bsp_halport.cpp
@@ -4,7 +4,7 @@ namespace Hardware
 {
     float MainFreq_MHz = 72.0f;
 
-    bool LogAtUart = false;
+    bool RTTLogAtUart = false;
 
     /**-----    配置CAN总线     -----**/
     CAN_HandleTypeDef* hcan_main = nullptr;
