+++ ReactorLibs/Bsps/bsp_system/bsp_halport.hpp
@@ -62,7 +62,7 @@ namespace Hardware
     extern float MainFreq_MHz;
 
     /// @note 如果本标志位被激活，日志会被同步发送到 UART_HOST
-    extern bool LogAtUart;
+    extern bool RTTLogAtUart;
     
     /***---------------     框架CAN    ---------------***/
     /// @brief 框架所用CAN句柄
