+++ ReactorLibs/Mods/inc/msg_coder.hpp
@@ -36,7 +36,7 @@ public:
     FrameStruct frame; 
 
     /// @brief 串口实例
-    BspUart_Instance uart_inst;
+    BSP::UART::Handler uart_inst;
 
     /// @brief 初始化函数
     void Init(UART_HandleTypeDef *huart);
